+++ CHANGELOG.md
@@ -11,6 +11,15 @@ Here we write upgrading notes for brands. It's a team effort to make them as
 straightforward as possible.
 
 ### Added
+
+### Changed
+
+### Fixed
+
+
+## 2023-04-26_e311fc1___2023-04-26_217f745
+
+### Added
 - USB drive dump progress
 - Ability to resume failed drive dump
 
@@ -28,6 +37,11 @@ straightforward as possible.
 - Dumping duplicate files
 - Drive dump failing after 1 minute mark
 - Adaptive uploads and users/rooms list for desctop
+- Show Cargo Room tab regardless of media settings
+- Error entering room on main drive fix
+- Markup for admin room fixed
+- Adaptive uploads and users/rooms list for mobile
+- handle lost messages when viewing gallery
 
 
 ## 2023-04-15_c77d709___2023-04-15_64148f8